# backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LocalMind API",
        "description": "OpenAPI documentation for the LocalMind backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Docker-served LocalMind backend"
        }
    ],
    "paths": {
...
        "/api/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "List questions",
                "operationId": "9bd3c1fee515c0270b73811bf64c312a",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Filter questions by keyword",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter questions by location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated question list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuestionPaginatedResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a new question",
                "operationId": "9ec74a56df8171af4fd5149a9364ce30",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuestionMutationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/questions/{question}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Show a question with its responses",
                "operationId": "632f4a3fa7b9126ec1ce318e8f184a0a",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuestionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Question not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Update a question",
                "operationId": "899bb44a229896a3cd8f916e7c4ac34c",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Question updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuestionMutationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Delete a question",
                "operationId": "ebf34d7c982b324544e705e1bd5be6e2",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthorSummary": {
                "required": [
                    "id",
                    "name",
                    "initial"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lahcen"
                    },
                    "initial": {
                        "type": "string",
                        "example": "L"
                    }
                },
                "type": "object"
            },
            "QuestionMetrics": {
                "required": [
                    "responses_count",
                    "favorites_count"
                ],
                "properties": {
                    "responses_count": {
                        "type": "integer",
                        "example": 4
                    },
                    "favorites_count": {
                        "type": "integer",
                        "example": 9
                    }
                },
                "type": "object"
            },
            "ResponseResource": {
                "required": [
                    "id",
                    "content",
                    "question_id",
                    "created_at",
                    "author",
                    "is_owner"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "content": {
                        "type": "string",
                        "example": "You can find a pharmacy near the train station."
                    },
                    "question_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2 minutes ago"
                    },
                    "author": {
                        "$ref": "#/components/schemas/AuthorSummary"
                    },
                    "is_owner": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "QuestionResource": {
                "required": [
                    "id",
                    "title",
                    "content",
                    "location",
                    "created_at",
                    "timestamp",
                    "author",
                    "metrics",
                    "responses"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "title": {
                        "type": "string",
                        "example": "Where can I find a late-night pharmacy?"
                    },
                    "content": {
                        "type": "string",
                        "example": "I need a pharmacy open after 10pm in Casablanca."
                    },
                    "location": {
                        "type": "string",
                        "example": "Casablanca"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "3 hours ago"
                    },
                    "timestamp": {
                        "type": "string",
                        "example": "10 Mar 2026 21:30"
                    },
                    "author": {
                        "$ref": "#/components/schemas/AuthorSummary"
                    },
                    "metrics": {
                        "$ref": "#/components/schemas/QuestionMetrics"
                    },
                    "responses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ResponseResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "QuestionMutationResponse": {
                "required": [
                    "message",
                    "data"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Question created successfully!"
                    },
                    "data": {
                        "$ref": "#/components/schemas/QuestionResource"
                    }
                },
                "type": "object"
            },
...
            "MessageResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operation completed successfully."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
...
            "QuestionRequest": {
                "required": [
                    "title",
                    "content",
                    "location"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Where can I repair my bike?",
                        "minLength": 5
                    },
                    "content": {
                        "type": "string",
                        "example": "Looking for a trusted bike repair shop near Maarif.",
                        "minLength": 10
                    },
                    "location": {
                        "type": "string",
                        "example": "Casablanca"
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions?page=1",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions?page=2",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMetaLink": {
                "required": [
                    "url",
                    "label",
                    "active"
                ],
                "properties": {
                    "url": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions?page=1",
                        "nullable": true
                    },
                    "label": {
                        "type": "string",
                        "example": "1"
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "required": [
                    "current_page",
                    "from",
                    "last_page",
                    "links",
                    "path",
                    "per_page",
                    "to",
                    "total"
                ],
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 3
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaginationMetaLink"
                        }
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost:8080/api/questions"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 31
                    }
                },
                "type": "object"
            },
            "QuestionPaginatedResponse": {
                "required": [
                    "data",
                    "links",
                    "meta"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuestionResource"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the bearer token returned by the register or login endpoint.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}